🍻function avatars, with two parameters, posters and users
Each forum post will include a list of user avatar images which represent all of the users participating in the conversation for that topic. find the correct user in the users array, const user = users.find(). customize the avatar's size, set a value of 30, The result of using the replace method in user.avatar_template